Unify Maya shot open_version_browser and save via _resolve_current_stream
